handling multi sig as fee/reward receiver (call) updating ref price is reward_receiver is not set

+++ contracts/interfaces/IRewardFeeReceiver.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.20;
 
 interface IRewardFeeReceiver {
-    function DepositFromVault(uint _amount) external;
+    function DepositFromVault(uint256 _amount) external;
 }
